feat(masthead-v2): create base megamenu header and adjust layout
