+++ packages/web-components/src/components/masthead/masthead.scss
@@ -756,10 +756,10 @@ $breakpoint--desktop-nav: 960px;
   }
 }
 
+:host(#{$dds-prefix}-megamenu-heading),
 :host(#{$dds-prefix}-megamenu-category-heading) {
   display: block;
   padding: 0 $spacing-05;
-  margin-bottom: $spacing-06;
   width: 66.66%;
 
   @include carbon--breakpoint('lg') {
@@ -767,11 +767,8 @@ $breakpoint--desktop-nav: 960px;
   }
 
   h2 {
-    @include carbon--type-style(heading-05, true);
-
     display: inline-block;
-    margin-bottom: $spacing-03;
-    color: $text-02;
+    color: $text-01;
 
     a {
       color: inherit;
@@ -791,9 +788,8 @@ $breakpoint--desktop-nav: 960px;
 
     svg {
       position: relative;
-      top: $spacing-02;
       margin: 0 $spacing-03;
-      fill: $icon-02;
+      fill: $icon-01;
       flex-shrink: 0;
 
       /* stylelint-disable-next-line comment-whitespace-inside */
@@ -819,10 +815,39 @@ $breakpoint--desktop-nav: 960px;
     @include carbon--type-style(body-short-01);
 
     display: block;
+    margin-top: $spacing-03;
     color: $text-02;
 
     @include carbon--breakpoint('lg') {
       width: 40%;
+    }
+
+    &[data-has-content] {
+      margin-bottom: $spacing-06;
+    }
+  }
+}
+
+:host(#{$dds-prefix}-megamenu-heading) {
+  margin-bottom: $spacing-06;
+
+  h2 {
+    @include carbon--type-style(heading-05, true);
+
+    svg {
+      top: $spacing-03;
+    }
+  }
+}
+
+:host(#{$dds-prefix}-megamenu-category-heading) {
+  margin-bottom: $spacing-03;
+
+  h2 {
+    @include carbon--type-style(heading-03, true);
+
+    svg {
+      top: $spacing-02;
     }
   }
 }
