+++ packages/web-components/src/components/masthead/megamenu-category-heading.ts
@@ -7,62 +7,17 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import { html, customElement, LitElement, property } from 'lit-element';
-import ArrowRight32 from 'carbon-web-components/es/icons/arrow--right/32.js';
-import ddsSettings from '../../internal/vendor/@carbon/ibmdotcom-utilities/utilities/settings/settings';
+import ArrowRight20 from 'carbon-web-components/es/icons/arrow--right/20.js';
 import styles from './masthead.scss';
-
-const { stablePrefix: ddsPrefix } = ddsSettings;
+import DDSMegaMenuHeading from './megamenu-heading';
 
 /**
  * MegaMenu Category Heading.
  *
  * @element dds-megamenu-category-heading
  */
-@customElement(`${ddsPrefix}-megamenu-category-heading`)
-class DDSMegaMenuCategoryHeading extends LitElement {
-  /**
-   * href.
-   */
-  @property({ reflect: true })
-  href = '';
-
-  /**
-   * Megamenu content category heading
-   */
-  @property({ reflect: true })
-  title = '';
-
-  /**
-   * Render category heading as link.
-   */
-  protected renderLink() {
-    return html`
-      <a href="${this.href}" data-attribute1="headerNav" data-attribute2="FlatHdline" data-attribute3="${this.title}">
-        ${this.title} ${ArrowRight32()}
-      </a>
-    `;
-  }
-
-  /**
-   * Render category heading as plain text.
-   */
-  protected renderPlain() {
-    return html`
-      ${this.title}
-    `;
-  }
-
-  render() {
-    const hasHref = Boolean(this.href);
-
-    return html`
-      <h2 ?data-has-href="${Boolean(hasHref)}">
-        ${hasHref ? this.renderLink() : this.renderPlain()}
-      </h2>
-      <span><slot></slot></span>
-    `;
-  }
+class DDSMegaMenuCategoryHeading extends DDSMegaMenuHeading {
+  protected _arrowIcon = ArrowRight20();
 
   static styles = styles; // `styles` here is a `CSSResult` generated by custom WebPack loader
 }
